Refactor Display code into its own package
